+++ src/builder.rs
@@ -2,10 +2,10 @@ use crate::{
     collection::{MenuItemCollection, MenuItems},
     interaction::{InputAdapterSource, InputState},
     selection_indicator::{style::IndicatorStyle, SelectionIndicatorController},
+    theme::Theme,
     Menu, MenuItem, MenuState, MenuStyle, NoItems,
 };
 use core::marker::PhantomData;
-use embedded_graphics::prelude::PixelColor;
 use embedded_layout::{
     layout::linear::LinearLayout,
     object_chain::ChainElement,
@@ -19,7 +19,7 @@ where
     IT: InputAdapterSource<R>,
     S: IndicatorStyle,
     P: SelectionIndicatorController,
-    C: PixelColor,
+    C: Theme,
 {
     title: T,
     items: LL,
@@ -32,7 +32,7 @@ where
     S: IndicatorStyle,
     IT: InputAdapterSource<R>,
     P: SelectionIndicatorController,
-    C: PixelColor,
+    C: Theme,
 {
     pub const fn new(title: T, style: MenuStyle<S, IT, P, R, C>) -> Self {
         Self {
@@ -48,8 +48,8 @@ where
     T: AsRef<str>,
     IT: InputAdapterSource<R>,
     P: SelectionIndicatorController,
-    S: IndicatorStyle,
-    C: PixelColor + Default + 'static,
+    S: IndicatorStyle<Color = C>,
+    C: Theme,
 {
     pub fn add_item<I: MenuItem<R>>(self, mut item: I) -> MenuBuilder<T, IT, Chain<I>, R, P, S, C> {
         item.set_style(&self.style);
@@ -88,8 +88,8 @@ where
     IT: InputAdapterSource<R>,
     Chain<CE>: MenuItemCollection<R>,
     P: SelectionIndicatorController,
-    S: IndicatorStyle,
-    C: PixelColor + Default + 'static,
+    S: IndicatorStyle<Color = C>,
+    C: Theme,
 {
     pub fn add_item<I: MenuItem<R>>(
         self,
@@ -132,8 +132,8 @@ where
     Link<I, CE>: MenuItemCollection<R> + ChainElement,
     CE: MenuItemCollection<R> + ChainElement,
     P: SelectionIndicatorController,
-    S: IndicatorStyle,
-    C: PixelColor + Default + 'static,
+    S: IndicatorStyle<Color = C>,
+    C: Theme,
 {
     pub fn add_item<I2: MenuItem<R>>(
         self,
@@ -176,7 +176,7 @@ where
     VG: ViewGroup + MenuItemCollection<R>,
     P: SelectionIndicatorController,
     S: IndicatorStyle,
-    C: PixelColor,
+    C: Theme,
 {
     pub fn build(self) -> Menu<T, IT, VG, R, P, S, C> {
         self.build_with_state(MenuState {